+++ R/cutData.R
@@ -7,35 +7,35 @@
 #' This section give a brief description of each of the define levels of `type`.
 #' Note that all time dependent types require a column `date`.
 #'
-#' - `"default"` does not split the data but will describe the levels as a date
+#'  - `"default"` does not split the data but will describe the levels as a date
 #' range in the format "day month year".
 #'
-#' - `"year"` splits the data by each year.
+#'  - `"year"` splits the data by each year.
 #'
-#' - `"month"` splits the data by month of the year.
+#'  - `"month"` splits the data by month of the year.
 #'
-#' - `"hour"` splits the data by hour of the day.
+#'  - `"hour"` splits the data by hour of the day.
 #'
-#' - `"monthyear"` splits the data by year and month. It differs from month in
+#'  - `"monthyear"` splits the data by year and month. It differs from month in
 #' that a level is defined for each month of the data set. This is useful
 #' sometimes to show an ordered sequence of months if the data set starts half
 #' way through a year; rather than starting in January.
 #'
-#' - `"weekend"` splits the data by weekday and weekend.
+#'  - `"weekend"` splits the data by weekday and weekend.
 #'
-#' - `"weekday"` splits the data by day of the week - ordered to start Monday.
+#'  - `"weekday"` splits the data by day of the week - ordered to start Monday.
 #'
-#' - `"season"` splits data up by season. In the northern hemisphere winter =
+#'  - `"season"` splits data up by season. In the northern hemisphere winter =
 #' December, January, February; spring = March, April, May etc. These
 #' definitions will change of `hemisphere = "southern"`.
 #'
-#' - `"seasonyear"` (or `"yearseason"`) will split the data into year-season
+#'  - `"seasonyear"` (or `"yearseason"`) will split the data into year-season
 #' intervals, keeping the months of a season together. For example, December
 #' 2010 is considered as part of winter 2011 (with January and February 2011).
 #' This makes it easier to consider contiguous seasons. In contrast, `type =
 #' "season"` will just split the data into four seasons regardless of the year.
 #'
-#' - `"daylight"` splits the data relative to estimated sunrise and sunset to
+#'  - `"daylight"` splits the data relative to estimated sunrise and sunset to
 #' give either daylight or nighttime. The cut is made by `cutDaylight` but more
 #' conveniently accessed via `cutData`, e.g. `cutData(mydata, type = "daylight",
 #' latitude = my.latitude, longitude = my.longitude)`. The daylight estimation,
@@ -45,11 +45,11 @@
 #' is based on NOAA methods. Measurement location should be set using `latitude`
 #' (+ to North; - to South) and `longitude` (+ to East; - to West).
 #'
-#' - `"dst"` will split the data by hours that are in daylight saving time (DST)
+#'  - `"dst"` will split the data by hours that are in daylight saving time (DST)
 #' and hours that are not for appropriate time zones. The option also requires
 #' that the local time zone is given e.g. `local.tz = "Europe/London"`,
 #' `local.tz = "America/New_York"`. Each of the two periods will be in
-#' *local time*. The main purpose of this option is to test whether there
+#'  *local time*. The main purpose of this option is to test whether there
 #' is a shift in the diurnal profile when DST and non-DST hours are compared.
 #' This option is particularly useful with the [timeVariation()] function. For
 #' example, close to the source of road vehicle emissions, "rush-hour" will tend
@@ -62,7 +62,7 @@
 #' whether the variation in a pollutant is driven by man-made emissions or
 #' natural processes.
 #'
-#' - `"wd"` splits the data by 8 wind sectors and requires a column `wd`: "NE",
+#'  - `"wd"` splits the data by 8 wind sectors and requires a column `wd`: "NE",
 #' "E", "SE", "S", "SW", "W", "NW", "N".
 #'
 #' Note that all the date-based types, e.g., `"month"`/`"year"` are derived from
@@ -79,9 +79,10 @@
 #'   name is supplied [cutData()] will split the data into four quantiles.
 #'   Factors levels will be used to split the data without any adjustment.
 #' @param names By default, the columns created by [cutData()] are named after
-#'   their `type` option. Specifying `names` defines other names for the columns,
-#'   which map onto the `type` options in the same order they are given. The
-#'   length of `names` should therefore be equal to the length of `type`.
+#'   their `type` option. Specifying `names` defines other names for the
+#'   columns, which map onto the `type` options in the same order they are
+#'   given. The length of `names` should therefore be equal to the length of
+#'   `type`.
 #' @param suffix If `name` is not specified, `suffix` will be appended to any
 #'   added columns that would otherwise overwrite existing columns. For example,
 #'   `cutData(mydata, "nox", suffix = "_cuts")` would append a `nox_cuts` column
@@ -104,14 +105,34 @@
 #' @param latitude,longitude The decimal latitude and longitudes used when `type
 #'   = "daylight"`. Note that locations west of Greenwich have negative
 #'   longitudes.
+#' @param drop How to handle empty factor levels. One of:
+#'
+#'  - `"default"`: Sensible defaults selected on a case-by-case basis for
+#'   different `type` options.
+#'
+#'  - `"empty"`: Drop all empty factor levels.
+#'
+#'  - `"none"`: Retain all empty factor levels, where possible. For example,
+#'   for `type = "hour"`, all factor levels from `0` and `23` will be
+#'   represented.
+#'
+#'  - `"outside"`: Retain empty factor levels within the range of the data.
+#'   For example, for `type = "hour"` when the data only contains data for 1 AM
+#'   and 5 AM, the factor levels, `1`, `2`, `3`, `4` and `5` will be retained.
+#'
+#'   Some of these options only apply to certain `type` options. For example,
+#'   for `type = "year"`, `"outside"` is equivalent to `"none"` as there is no
+#'   fixed range of years to use in the `"none"` case.
+#'
 #' @param ... All additional parameters are passed on to next function(s).
 #' @export
 #' @return Returns the data frame, `x`, with columns appended as defined by
 #'   `type` and `name`.
 #' @author David Carslaw
+#' @author Jack Davison
 #' @author Karl Ropkins (`"daylight"` option)
 #' @examples
-#' ## split data by day of the week
+#' # split data by day of the week
 #' mydata <- cutData(mydata, type = "weekday")
 #' names(mydata)
 #' head(mydata)
@@ -127,8 +148,11 @@ cutData <- function(
   local.tz = NULL,
   latitude = 51,
   longitude = -0.5,
+  drop = c("default", "empty", "outside", "none"),
   ...
 ) {
+  drop <- rlang::arg_match(drop, c("default", "empty", "outside", "none"))
+
   if (!is.null(names)) {
     if (length(names) != length(type)) {
       cli::cli_abort(
@@ -227,53 +251,62 @@ cutData <- function(
     }
 
     if (type == "year") {
-      x[[name]] <- cutVecYear(x$date)
+      x[[name]] <- cutVecYear(x$date, drop = drop)
     }
 
     if (type == "hour") {
-      x[[name]] <- cutVecHour(x$date)
+      x[[name]] <- cutVecHour(x$date, drop = drop)
     }
 
     if (type == "month") {
-      x[[name]] <- cutVecMonth(x$date, is.axis = is.axis)
+      x[[name]] <- cutVecMonth(x$date, is.axis = is.axis, drop = drop)
     }
 
     if (type %in% c("monthyear", "yearmonth")) {
-      x[[name]] <- cutVecMonthyear(x$date, is.axis = is.axis)
+      x[[name]] <- cutVecMonthyear(x$date, is.axis = is.axis, drop = drop)
     }
 
     if (type == "week") {
-      x[[name]] <- cutVecWeek(x$date)
+      x[[name]] <- cutVecWeek(x$date, drop = drop)
     }
 
     if (type == "season") {
-      x[[name]] <- cutVecSeason(x$date, hemisphere = hemisphere)
+      x[[name]] <- cutVecSeason(x$date, hemisphere = hemisphere, drop = drop)
     }
 
     if (type %in% c("seasonyear", "yearseason")) {
-      x[[name]] <- cutVecSeasonyear(x$date, hemisphere = hemisphere)
+      x[[name]] <- cutVecSeasonyear(
+        x$date,
+        hemisphere = hemisphere,
+        drop = drop
+      )
     }
 
     if (type == "weekend") {
-      x[[name]] <- cutVecWeekend(x$date)
+      x[[name]] <- cutVecWeekend(x$date, drop = drop)
     }
 
     if (type == "weekday") {
-      x[[name]] <- cutVecWeekday(x$date, start.day = start.day)
+      x[[name]] <- cutVecWeekday(
+        x$date,
+        is.axis = is.axis,
+        start.day = start.day,
+        drop = drop
+      )
     }
 
     if (type == "wd") {
       x <- dropNAbyType(x, "wd")
-      x[[name]] <- cutVecWinddir(x$wd)
+      x[[name]] <- cutVecWinddir(x$wd, drop = drop)
     }
 
     if (type %in% c("dst", "bstgmt", "gmtbst")) {
       type <- "dst" ## keep it simple
-      x[[name]] <- cutVecDST(x$date, local.tz = local.tz)
+      x[[name]] <- cutVecDST(x$date, local.tz = local.tz, drop = drop)
     }
 
     if (type == "daylight") {
-      x[[name]] <- cutVecDaylight(x$date, latitude, longitude, ...)
+      x[[name]] <- cutVecDaylight(x$date, latitude, longitude, ..., drop = drop)
     }
 
     return(x)
@@ -285,8 +318,8 @@ cutData <- function(
   return(x)
 }
 
-#' Drop missing values and warn that it has happened
-#' @noRd
+#'  Drop missing values and warn that it has happened
+#'  @noRd
 dropNAbyType <- function(x, type) {
   if (anyNA(x[[type]])) {
     lenNA <- length(which(is.na(x[[type]])))
@@ -298,8 +331,8 @@ dropNAbyType <- function(x, type) {
   return(x)
 }
 
-#' Cut a numeric vector into quantiles
-#' @noRd
+#'  Cut a numeric vector into quantiles
+#'  @noRd
 cutVecNumeric <- function(x, type, n.levels, is.axis) {
   temp.levels <-
     levels(cut(
@@ -334,86 +367,193 @@ cutVecNumeric <- function(x, type, n.levels, is.axis) {
   return(x)
 }
 
-#' Cut a vector into a 'year' factor
-#' @noRd
-cutVecYear <- function(x) {
-  ordered(lubridate::year(x))
+#'  Cut a vector into a 'year' factor
+#'  @noRd
+cutVecYear <- function(x, drop) {
+  x <- lubridate::year(x)
+  if (drop %in% c("default", "none", "outside")) {
+    x <- ordered(x)
+  } else {
+    x_range <- range(unique(x), na.rm = TRUE)
+    levels <- seq(x_range[1], x_range[2], by = 1L)
+    x <- ordered(x, levels = levels)
+  }
+  x
 }
 
-#' Cut a vector into a 'hour' factor
-#' @noRd
-cutVecHour <- function(x) {
-  ordered(lubridate::hour(x))
+#'  Cut a vector into a 'hour' factor
+#'  @noRd
+cutVecHour <- function(x, drop) {
+  x <- lubridate::hour(x)
+  if (drop %in% c("none")) {
+    x <- ordered(x, levels = 0:23)
+  } else if (drop == "outside") {
+    x_range <- range(unique(x), na.rm = TRUE)
+    levels <- seq(x_range[1], x_range[2], by = 1L)
+    x <- ordered(x, levels = levels)
+  } else if (drop %in% c("default", "empty")) {
+    x <- ordered(x)
+  }
+  x
 }
 
-#' Cut a vector into a factor of weeks of the year
-#' @noRd
-cutVecWeek <- function(x) {
-  x <- format(x, "%W")
-  x <- ordered(x, levels = unique(x))
+#'  Cut a vector into a factor of weeks of the year
+#'  @noRd
+cutVecWeek <- function(x, drop) {
+  x <- as.numeric(format(x, "%W"))
+  if (drop %in% c("none")) {
+    x <- ordered(two(x), levels = two(0:53))
+  } else if (drop == "outside") {
+    x_range <- range(unique(x), na.rm = TRUE)
+    levels <- seq(x_range[1], x_range[2], by = 1L)
+    x <- ordered(two(x), levels = two(levels))
+  } else if (drop %in% c("default", "empty")) {
+    x <- ordered(two(x), levels = two(unique(x)))
+  }
   return(x)
 }
 
-#' Cut a date vector into weekday/weekend
-#' @noRd
-cutVecWeekend <- function(x) {
+#'  Cut a date vector into weekday/weekend
+#'  @noRd
+cutVecWeekend <- function(x, drop) {
   wdays <- lubridate::wday(x, week_start = 1L)
   x <- dplyr::case_match(wdays, 1:5 ~ "weekday", 6:7 ~ "weekend")
-  x <- ordered(x, levels = c("weekday", "weekend"))
-  return(x)
-}
-
-#' Cut a date vector into weekdays
-#' @noRd
-cutVecWeekday <- function(x, start.day) {
-  if (start.day == 0L) {
-    start.day <- 7L
+  unique_vals <- unique(x)
+  if (drop %in% c("default", "none")) {
+    levels <- c("weekday", "weekend")
+  } else {
+    levels <- unique_vals
   }
-  x <- lubridate::wday(x, label = TRUE, abbr = FALSE, week_start = start.day)
-  levels <- levels(x)
-  levels <- levels[levels %in% x]
   x <- ordered(x, levels = levels)
   return(x)
 }
 
-#' Cut a vector into an ordered 'month' factor with no empty levels
-#' @noRd
-cutVecMonth <- function(x, is.axis) {
-  x <- lubridate::month(x, label = TRUE, abbr = is.axis)
-  levs <- levels(x)[levels(x) %in% x]
-  x <- factor(x, levels = levs, ordered = TRUE)
+#'  Cut a date vector into weekdays
+#'  @noRd
+cutVecWeekday <- function(x, is.axis, start.day, drop) {
+  if (start.day == 0L) {
+    start.day <- 7L
+  }
+  x <- lubridate::wday(x, label = TRUE, abbr = is.axis, week_start = start.day)
+  levels <- levels(x)
+  if (drop %in% c("default", "empty")) {
+    levels <- levels[levels %in% x]
+  } else if (drop == "none") {
+    levels <- levels
+  } else if (drop == "outside") {
+    levels
+    x_range <- range(unique(as.numeric(x)))
+    levels_int <- seq(x_range[1], x_range[2], by = 1L)
+    levels <- levels[levels_int]
+  }
+  x <- ordered(x, levels = levels)
   return(x)
 }
 
-#' Cut a vector into 'monthyears' (e.g., January 2020)
-#' @noRd
-cutVecMonthyear <- function(x, is.axis) {
+#'  Cut a vector into an ordered 'month' factor
+#'  @noRd
+cutVecMonth <- function(x, is.axis, drop) {
+  x <- lubridate::month(x, label = TRUE, abbr = is.axis)
+  levels <- levels(x)
+  if (drop %in% c("default", "empty")) {
+    levels <- levels[levels %in% x]
+  } else if (drop == "none") {
+    levels <- levels
+  } else if (drop == "outside") {
+    x_range <- range(unique(as.numeric(x)))
+    levels_int <- seq(x_range[1], x_range[2], by = 1L)
+    levels <- levels[levels_int]
+  }
+  x <- ordered(x, levels = levels)
+  return(x)
+}
+
+#'  Cut a vector into 'monthyears' (e.g., January 2020)
+#'  @noRd
+cutVecMonthyear <- function(x, is.axis, drop) {
   str <- "%B %Y"
   if (is.axis) {
     str <- "%b %Y"
   }
 
-  x <- format(x, str)
-  x <- ordered(x, levels = unique(x))
+  # get years and months
+  yrs <- cutVecYear(x, drop = "none")
+  mnths <- cutVecMonth(x, is.axis = is.axis, drop = "none")
+
+  # get combinations of years and months
+  levels <-
+    tidyr::crossing(
+      year = factor(levels(yrs), levels = levels(yrs)),
+      month = factor(levels(mnths), levels = levels(mnths))
+    ) |>
+    dplyr::mutate(
+      level = paste(.data$month, .data$year, sep = " ")
+    ) |>
+    dplyr::pull(.data$level)
+
+  # combine actual years and months
+  x <- paste(mnths, yrs, sep = " ")
+
+  # get the factor levels
+  if (drop %in% c("default", "empty")) {
+    levels <- unique(x)
+  } else if (drop == "none") {
+    levels <- levels
+  } else if (drop == "outside") {
+    start <- which(levels == dplyr::first(x))
+    end <- which(levels == dplyr::last(x))
+    levels <- levels[start:end]
+  }
+
+  x <- ordered(x, levels = levels)
+
   return(x)
 }
 
-#' Cut into season year (e.g., Summer 2020)
-#' @noRd
-cutVecSeasonyear <- function(x, hemisphere) {
-  seasons <- cutVecSeason(x, hemisphere = hemisphere)
+#'  Cut into season year (e.g., Summer 2020)
+#'  @noRd
+cutVecSeasonyear <- function(x, hemisphere, drop) {
+  # get seasons/years
+  seasons <- cutVecSeason(x, hemisphere = hemisphere, drop = "none")
   years <- lubridate::year(x)
+
+  # adjust year if month is 12 - belongs to "winter" next year
   months <- lubridate::month(x)
   years[months == 12] <- years[months == 12] + 1L
 
-  x <- paste(seasons, "-", years)
-  x <- ordered(x, levels = unique(x))
+  # get combinations of years and months
+  levels <-
+    tidyr::crossing(
+      year = seq(min(years, na.rm = TRUE), max(years, na.rm = TRUE), by = 1L),
+      season = factor(levels(seasons), levels = levels(seasons))
+    ) |>
+    dplyr::mutate(
+      level = paste(.data$season, .data$year, sep = "-")
+    ) |>
+    dplyr::pull(.data$level)
+
+  # combine actual years and months
+  x <- paste(seasons, years, sep = "-")
+
+  # get the factor levels
+  if (drop %in% c("default", "empty")) {
+    levels <- unique(x)
+  } else if (drop == "none") {
+    levels <- levels
+  } else if (drop == "outside") {
+    start <- which(levels == dplyr::first(x))
+    end <- which(levels == dplyr::last(x))
+    levels <- levels[start:end]
+  }
+
+  x <- ordered(x, levels = levels)
+
   return(x)
 }
 
-#' Cut wind direction into bins
-#' @noRd
-cutVecWinddir <- function(x) {
+#'  Cut wind direction into bins
+#'  @noRd
+cutVecWinddir <- function(x, drop) {
   x <- cut(
     x,
     breaks = seq(22.5, 382.5, 45),
@@ -422,17 +562,29 @@ cutVecWinddir <- function(x) {
 
   x[is.na(x)] <- "N" # for wd < 22.5
 
-  x <- ordered(x, levels = c("N", "NE", "E", "SE", "S", "SW", "W", "NW"))
+  levels <- c("N", "NE", "E", "SE", "S", "SW", "W", "NW")
+
+  if (drop %in% c("default", "none")) {
+    levels <- levels
+  } else if (drop == "empty") {
+    levels <- levels[levels %in% x]
+  } else if (drop == "outside") {
+    start <- which(levels == dplyr::first(x))
+    end <- which(levels == dplyr::last(x))
+    levels <- levels[start:end]
+  }
+
+  x <- ordered(x, levels = levels)
 
   return(x)
 }
 
-#' Cut dates into DST
-#' @noRd
-cutVecDST <- function(x, local.tz) {
+#'  Cut dates into DST
+#'  @noRd
+cutVecDST <- function(x, local.tz, drop) {
   ## how to extract BST/GMT
   if (is.null(local.tz)) {
-    message("missing time zone, assuming Europe/London")
+    cli::cli_warn("missing time zone, assuming Europe/London")
     local.tz <- "Europe/London"
   }
 
@@ -448,26 +600,31 @@ cutVecDST <- function(x, local.tz) {
 
   x <- dplyr::case_match(isdst, 0 ~ "Non-DST", 1 ~ "DST")
 
-  x <- factor(x, levels = sort(unique(x)))
+  if (drop %in% c("default", "empty", "outside")) {
+    x <- factor(x, levels = unique(x))
+  } else if (drop == "none") {
+    x <- factor(x, levels = c("DST", "Non-DST"))
+  }
 
   return(x)
 }
 
-#' Cut date vector into daylight or not
+#'  Cut date vector into daylight or not
 #'
-#' calculations use (lat, long) position relative to sun to estimate if daylight
-#' or nighttime hour solar.noon.lst, etc are factions of day seconds into that
-#' day = p.time * 86400 so for example sunset time is as.POSIXct(sunset.time.lst
-#' * 86400, origin = format(x$date, "%Y-%m-%d")) (assuming you do not run into
-#' next day!) currently unsure about extremes long nights and days at poles need
-#' checking
+#'  calculations use (lat, long) position relative to sun to estimate if daylight
+#'  or nighttime hour solar.noon.lst, etc are factions of day seconds into that
+#'  day = p.time * 86400 so for example sunset time is as.POSIXct(sunset.time.lst
+#'  * 86400, origin = format(x$date, "%Y-%m-%d")) (assuming you do not run into
+#'  next day!) currently unsure about extremes long nights and days at poles need
+#'  checking
 #'
-#' @noRd
+#'  @noRd
 cutVecDaylight <- function(
   x,
   latitude = 51.522393,
   longitude = -0.154700,
-  ...
+  ...,
+  drop
 ) {
   # back-compatibility
   x <- data.frame(date = x)
@@ -631,25 +788,31 @@ cutVecDaylight <- function(
       )
     )
   )
-  # as ordered factor
-  daylight <-
-    factor(
-      daylight,
-      levels = c(TRUE, FALSE),
-      labels = c("daylight", "nighttime")
-    )
 
-  return(daylight)
+  x <- dplyr::if_else(
+    daylight,
+    "daylight",
+    false = "nighttime",
+    missing = NA
+  )
+
+  if (drop %in% c("default", "empty", "outside")) {
+    x <- factor(x, levels = unique(x))
+  } else if (drop == "none") {
+    x <- factor(x, levels = c("daylight", "nighttime"))
+  }
+
+  return(x)
 }
 
-#' Cut a vector into seasons
-#' @noRd
-cutVecSeason <- function(x, hemisphere) {
+#'  Cut a vector into seasons
+#'  @noRd
+cutVecSeason <- function(x, hemisphere, drop) {
   hemisphere <- rlang::arg_match(hemisphere, c("northern", "southern"))
 
   # need to work out month names local to the user and extract first letter
   month_names_local <-
-    cutVecMonth(ISOdate(2000, 1:12, 1), is.axis = FALSE) |>
+    cutVecMonth(ISOdate(2000, 1:12, 1), is.axis = FALSE, drop = "none") |>
     substr(1, 1)
 
   # Function to create, e.g., 'winter (JFM)'
@@ -671,7 +834,7 @@ cutVecSeason <- function(x, hemisphere) {
         c(9, 10, 11) ~ make_season_name("autumn", c(9, 10, 11))
       )
 
-    seasons <-
+    levels <-
       c(
         make_season_name("spring", c(3, 4, 5)),
         make_season_name("summer", c(6, 7, 8)),
@@ -679,9 +842,13 @@ cutVecSeason <- function(x, hemisphere) {
         make_season_name("winter", c(12, 1, 2))
       )
 
-    seasons <- seasons[seasons %in% x]
+    if (drop %in% c("default", "empty")) {
+      levels <- levels[levels %in% x]
+    } else if (drop %in% c("none", "outside")) {
+      levels <- levels
+    }
 
-    x <- ordered(x, levels = seasons)
+    x <- ordered(x, levels = levels)
   } else {
     x <-
       dplyr::case_match(
@@ -692,7 +859,7 @@ cutVecSeason <- function(x, hemisphere) {
         c(9, 10, 11) ~ make_season_name("spring", c(9, 10, 11))
       )
 
-    seasons <-
+    levels <-
       c(
         make_season_name("spring", c(9, 10, 11)),
         make_season_name("summer", c(12, 1, 2)),
@@ -700,9 +867,13 @@ cutVecSeason <- function(x, hemisphere) {
         make_season_name("winter", c(6, 7, 8))
       )
 
-    seasons <- seasons[seasons %in% x]
+    if (drop %in% c("default", "empty")) {
+      levels <- levels[levels %in% x]
+    } else if (drop %in% c("none", "outside")) {
+      levels <- levels
+    }
 
-    x <- ordered(x, levels = seasons)
+    x <- ordered(x, levels = levels)
   }
 
   return(x)